rename ae_*.c to ae_*.h
